+++ src/index.tsx
@@ -1,5 +1,5 @@
 // import meta from "./meta";
-import { _createConcepts as createConcepts } from "./createConcepts";
+import { _createStoreHelpers as createStoreHelpers } from "./create";
 
 export type InitialItemsState<
   T_defaultStateFunctionType extends (...args: any) => any
@@ -7,15 +7,19 @@ export type InitialItemsState<
   [itemName: string]: ReturnType<T_defaultStateFunctionType>; // : AtLeastOne<T>;
 };
 
-export type ConceptsHelperTypes<
+export type StoreHelperTypes<
   T_GetState extends () => { [key: string]: { [key: string]: any } },
   T_GetRefs extends () => { [key: string]: { [key: string]: any } },
   T_ItemType extends keyof ReturnType<T_GetState> & keyof ReturnType<T_GetRefs>
 > = {
   ItemType: keyof ReturnType<T_GetState>;
   AllItemsState: ReturnType<T_GetState>[T_ItemType];
-  ItemState: ReturnType<T_GetState>[T_ItemType][keyof ReturnType<T_GetState>[T_ItemType]];
-  ItemRefs: ReturnType<T_GetRefs>[T_ItemType][keyof ReturnType<T_GetRefs>[T_ItemType]];
+  ItemState: ReturnType<T_GetState>[T_ItemType][keyof ReturnType<
+    T_GetState
+  >[T_ItemType]];
+  ItemRefs: ReturnType<T_GetRefs>[T_ItemType][keyof ReturnType<
+    T_GetRefs
+  >[T_ItemType]];
 };
 
 // for generating items with names
@@ -37,4 +41,4 @@ export function makeInitialState({
   return newInitialState;
 }
 
-export { createConcepts };
+export { createStoreHelpers };